apps/backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
supabase
psycopg2-binary
pandas
numpy
lxml
httpx
python-multipart
tqdm
google-genai
Pillow
python-multipart